requirements.txt
pydantic>=2.0
neo4j>=5.0
httpx>=0.27
rich>=13.0
python-dateutil>=2.9
